Update Azure Spot Termination checking
* Query the Azure metadata service for scheduled termination
events and decode the response JSON. In the past it may have
been sufficient to look for OK/200 responses as the indicator,
but now it seems that the default is a JSON response with zero
events.
* When any scheduled termination event is observed, start the
stop procedures (e.g. drain and/or delete, depending)

Docs: https://learn.microsoft.com/en-us/azure/virtual-machine-scale-sets/virtual-machine-scale-sets-terminate-notification#get-terminate-notifications